sales_visit/requirement.txt
django==2.2
djangorestframework==3.11
psycopg2==2.8
requests